backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
base58==2.1.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.6.2
cryptography==46.0.1
dateparser==1.2.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.12
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.0
idna==3.10
jsonalias==0.1.1
multidict==6.4.3
numpy==2.2.5
pandas==2.2.3
pillow==11.3.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.4
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
solana==0.36.7
solders==0.26.0
SQLAlchemy==2.0.40
starlette==0.46.2
ta==0.11.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
stellar-sdk
